minor fixes

+++ models/adversarial/lsb_steganography.py
@@ -108,7 +108,7 @@ class LsbDetection(Eve):
             covers, secrets = load_image_covers_and_ascii_bit_secrets(iterations_per_epoch*batch_size,
                                                                       scale=1./255.)
 
-            print((covers*255)%2)
+            print(((covers*255)%2).astype(int))
 
             # hidden_secrets = LsbSteganography.encode(covers, secrets)
